+++ core/deployment/operational_bootstrap_live/__init__.py
@@ -1,6 +1,7 @@
 """Dedicated controlled operational coordinator entrypoint."""
 
 from .artifacts import *
+from .acknowledgements import *
 from .coordinator import *
 from .composition import *
 from .models import *